requirements.txt
# --- Core Dependencies (Cross-Platform) ---
gymnasium==0.29.1
neat-python==0.92
pygame==2.5.2
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
graphviz>=0.20.1

# --- Physics Engine Dependencies ---
# Gymnasium's Box2D requires SWIG to compile the C++ physics engine.
# We use environment markers to ensure the correct handling per OS.

# Windows
swig==4.1.1.post1 ; sys_platform == 'win32'
box2d-py==2.3.5 ; sys_platform == 'win32'

# macOS
swig==4.1.1.post1 ; sys_platform == 'darwin'
box2d-py==2.3.5 ; sys_platform == 'darwin'